Move frame to v_frame
This commit includes:
- Moving frame's struct and impl to v_frame
- Introduction of LUMA_PADDING as a constant
- Introduction of multiple Public Trait for Frame Alocation,
  Calculating padding, new Tile of a frame
- Introduction of new video frame with padding constant as parameter
